-Fixed snow having undefined value in most biomes, causing to snow in incorrect biomes in Android -Fixed desert, ice desert and hell biomes having rain -Fixed slime not "squishing"

+++ source/world/entity/Slime.hpp
@@ -24,8 +24,8 @@ public:
 	bool canSpawn() override;
 
 public:
-	int m_oSquish;
-	int m_squish;
+	float m_oSquish;
+	float m_squish;
 	int m_size;
 
 private:
